+++ CHANGELOG.md
@@ -23,6 +23,7 @@ limitations under the License.
 - feat: module inspect considers lists and dicts for Torch module containers
 - feat: add JIT inspect
 - feat: add support for forward hooks for AOT and JIT tuning
+- fix: prevent cache dir override when there are two similar modules in jit tuning
 
 ## 0.1.0